+++ packages/patrol_finders/test/patrol_finder_test.dart
@@ -565,6 +565,36 @@ void main() {
         await $('done').waitUntilVisible();
         expect($('done').visible, true);
       });
+
+    
+
+      patrolWidgetTest('waits until widget is only visible at the topCenter alignment', ($) async {
+        await $.pumpWidget(
+          MaterialApp(
+            home: FutureBuilder(
+              future: Future<void>.delayed(const Duration(seconds: 3)),
+              builder: (context, snapshot) {
+                if (snapshot.connectionState == ConnectionState.done) {
+                  return const Column(children: [
+                    Text('some text'),
+                    SizedBox(height: 60),
+                    Text('some other text'),
+                  ],);
+                } else {
+                  return const Text('in progress');
+                }
+              },
+            ),
+          ),
+        );
+        await $(Column).waitUntilVisible(alignment: Alignment.topCenter);
+        expect($(Column).visible, false);
+        expect($(Column).isVisibleAt(alignment: Alignment.topCenter), true);
+        await expectLater(
+            $(Column).waitUntilVisible,
+            throwsA(isA<WaitUntilVisibleTimeoutException>()),
+          );
+      });
     });
 
     group('scrollTo()', () {